package/erlang: drop leftover comment
Commit b574a9606e62 (package/erlang: do not hard-code the Erlang
Interface Version (EI_VSN)) removed the need to care about keeping
the erlang version and its API version in sync, as the latter is
automatically extracted from installed files. The corresponding comment
was removed in that commit:

    # Whenever updating Erlang, this value should be updated as well, to the
    # value of EI_VSN in the file lib/erl_interface/vsn.mk

However, the removal of the comment above the package version was
leftover (probably because of a conflict with a version bump that was
improperly resolved).

Drop the comment now.

Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>