package/petitboot: prefer kexec-lite on powerpc
This is a better choice on platforms where it is available, so default
to it. Still select kexec for platforms where kexec-lite is not
available.

While we're at it, make use of the new
BR2_PACKAGE_KEXEC_LITE_ARCH_SUPPORTS symbol instead of relying of the
kexec-lite arch depends being a subject of kexec.

While we're at it, sort all run-time dependencies alphabetically and
don't repeat the comment.

Signed-off-by: Reza Arbab <arbab@linux.ibm.com>
[Arnout:
 - use BR2_PACKAGE_KEXEC_LITE_ARCH_SUPPORTS;
 - still allow full kexec by turning around the dependency;
 - extend commit message.
]
Signed-off-by: Arnout Vandecappelle <arnout@mind.be>