package/skopeo: new host package
skopeo is a command line utility that performs various operations on
container images and image repositories.

We introduce it as a host-only package, as the expected usage is to
transform an OCI image into a docker image, or to help in uploading an
OCI image to a (local) docker daemon or to a docker registry, for
example, in a post-build script.

As such, it needs a prompt in the host packages section.

Signed-off-by: Yann E. MORIN <yann.morin@orange.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>