package/flutter-engine: don't pollute user's home directory
flutter-engine will forcibly create ~/.dart/ and ~/.flutter/ and store
stuff in there.

This is however quite dirty and pollutes the user's home, and will also
not work when the home is read-only (e.g. shared in a container).

Forcibly redirect flutter-engine to the same location where we
redirected host-flutter-sdk-bin, using the same trick of redefining
HOME as for host-flutter-sdk-bin.

Ideally, we'd like to have some way to share this with all flutter-based
packages, but we so far have only a few of them, so it is too early to
even think about some commonalities (even less so about a shared infra).
So we just duplicate the setting for now, this can be revisited later.

Signed-off-by: Adam Duskett <adam.duskett@amarulasolutions.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>